[FIX] start recyclerview from index 1 in HomeFragment CategoryView (#147)
